+++ tests/conftest.py
@@ -85,3 +85,24 @@ def flint_containers(tmp_path_factory) -> Path:
     assert all(path.exists() for path in container_paths)
 
     return flint_container_path
+
+
+@pytest.fixture
+def ms_example_with_name(tmpdir):
+    def _ms_example(output_name: str):
+        ms_zip = Path(
+            get_packaged_resource_path(
+                package="flint.data.tests",
+                filename="SB39400.RACS_0635-31.beam0.small.ms.zip",
+            )
+        )
+        outpath = Path(tmpdir) / output_name
+        if outpath.exists():
+            message = f"{outpath=} already exists. Provide unique {output_name=}"
+            raise FileExistsError(message)
+
+        shutil.unpack_archive(ms_zip, outpath)
+
+        return Path(outpath) / "SB39400.RACS_0635-31.beam0.small.ms"
+
+    return _ms_example
