+++ tests/test_crystalball.py
@@ -15,12 +15,6 @@ from flint.predict.crystalball import (
 )
 from flint.utils import get_packaged_resource_path
 
-from .test_helpers import ms_example
-
-# This is a dummy usage to avoid the import of
-# the fixture being removed
-_ = ms_example
-
 
 def test_get_parser():
     """A simple test to obtain the parser. Tests imports as well."""
@@ -46,9 +40,9 @@ def test_expect_asserts():
         crystalball_predict(ms=ms, crystalball_options=crystalball_options)
 
 
-def test_example_prediction_ms(ms_example):
+def test_example_prediction_ms(ms_example_with_name):
     """Attempt to predict into the MS with crystalball"""
-    ms_path = ms_example("Jack")
+    ms_path = ms_example_with_name("Jack")
     assert ms_path.exists()
     ms = MS(path=ms_path)
     assert ms.model_column is None
