+++ tests/test_flows.py
@@ -4,12 +4,10 @@ import pytest
 
 from flint.logging import logger
 
-from .test_helpers import which
-
-if which("singularity") is None:
-    pytest.skip("Singularity is not installed", allow_module_level=True)
+from .conftest import which
 
 
+@pytest.mark.require_singularity
 def test_singularity():
     which_singularity = which("singularity")
     logger.info(f"Singularity is installed at: {which_singularity}")
